refactor(css): emit the math arithmetic beside the descriptors that bind it
The primitives were a module of their own for one reason — `lib/css/data.js`
cannot require what requires it. Emitting them from the generator instead puts
what each function does and the arithmetic it does it with in one file, and
drops the extra module. `lib/css/syntax.js` is unchanged in what it names:
nothing.

Generation now checks the descriptors against the emitted names rather than a
module's exports, so a name nothing defines still fails there.
