+++ lib/css/syntax.js
@@ -33,9 +33,11 @@ const {
 	SUBSTITUTION_FUNCTIONS,
 	UNIT_CONVERSION_TARGETS,
 	UNIT_GROUP_BASE,
-	ZERO_UNIT_KEEPING_PROPERTIES
+	ZERO_UNIT_KEEPING_PROPERTIES,
+	exactAdd,
+	exactDivide,
+	exactMultiply
 } = require("./data");
-const { exactAdd, exactDivide, exactMultiply } = require("./mathPrimitives");
 
 // spec: https://drafts.csswg.org/css-syntax/
